auth_service/requirements.txt
wheel==0.42.0
Flask==3.0.0
flask_sqlalchemy==3.1.1
kafka-python==2.0.2
pymongo==4.6.1
python-dotenv==1.0.0
mysqlclient==2.2.1
kafka==1.3.5
confluent_kafka==2.3.0
requests==2.31.0
prometheus-flask-exporter==0.23.0
flask-cors==4.0.0
PyJWT==2.8.0
prometheus-client==0.19.0
prometheus-flask-exporter==0.23.0